feature: import images for beaches, entertainment and culture categories
Import images for these three categories inside the drawable resource folder. These images will be used when creating lists and when opening the details of a list element.
Change also one of the entertainment entries since it is not active anymore. Then, add another entry in the beach category.

Integrates #1