feature: develop the welcome screen and adjust layouts
Let the main cards be elevated, in all the screens.
Develop the UI of the WelcomeScreen composable as it is better for starting the experience.
Import then the main background that will be used throughout the app as an artistical background.
Delete then all the descriptions related to the categories, and edit consequently the Category data class.

Integrates #9