feature: import dependencies for width size, create ui state and condition the ui
Provide ui conditions based on the input window width size provided. For now deal only with the compact one.
Define the UI state that is going to be used as the variable governing UI updates and recompositions.
Declare a ViewModel class for dealing with business logic.
Import the necessary dependencies for calculating the input window width size.

Integrates #9