feature: build app bar for compact layout and handle back and up navigation
Import the logo of the city for showing it in the WelcomeScreen.
Put all the composables for the compact layout in the MyCityCompact.kt file.
For the compact layout, build the app bar for the categories screen, entries screen and details screen. Handle the up navigation with the arrow button and the back navigation by overriding the back button behaviour.
Insert the necessary string resources and Typography for building the app bar of the application.

Integrates #9