feature: add layout for Medium devices
In the details screen, add the card composable that contains the rate of the entry.
Insert in the drawable resource folder the new image for entry 6 of accommodation.
Add in the MyCityApp composable the layout implementation for the medium device, and insert the necessary composables in the MyCityMedium.kt file.
Add also the necessary string resources in the strings.xml file.

Integrates #11