feature: adjust screen layout based on the current width and height
Pass the widthSizeClass inside the CategoryEntriesScreen and the DetailsScreen in order to perform a conditional layout, based on the current width of the device.
Insert then scrolling behaviours in the layouts that do not have it yet.
Change the weight of the views inside the navigation drawer content in expanded mode.

Integrates #11